requirements.txt
streamlit>=1.27.0
Pillow>=10.0.0
toml>=0.10.2
python-dotenv>=1.0.0
requests>=2.31.0
openai>=1.0.0
base64io>=1.0.3
